Improve highlighting, update to 0.3
This commit updates the syntax to match the one of the current stable
version of Nickel.

Doing so, we also make the highlighting group slighty more natural: we
distinguish between statements-like keyword and metadata keyword (which
are highlighted in a different way), highlight builtins like functions,
and fixed a few regular expressions.